reafcator(useUser, pages session and stats): create customize hook to get user and use it in pages session and stats
