+++ app/stats/page.jsx
@@ -1,9 +1,15 @@
-import React from 'react'
+"use client";
 
+import { useState } from "react";
+import { useSession } from "next-auth/react";
+import useUser from "@modules/client/userRequests/useUser";
+
+// Get all previous exercises stats by exercises id and uer id
 function Stats() {
-  return (
-    <div>Stats</div>
-  )
+  const { data: userSession, status } = useSession();
+  const { userId } = useUser(userSession);
+
+  
 }
 
-export default Stats
+export default Stats;
